Breaking apart tests into currency files
